Fix errors: unused and bugs linters report
